+++ CHANGELOG.md
@@ -8,6 +8,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ## [26.x.x]
 ### Changed
 - implement item navigation when showing details in no-split mode
+- add swipe down event to refresh item list for mobile devices
 
 ### Fixed
 - style fixes for `<blockquote>`, `<pre>` and `<code>` in articles